third_party/WebKit/LayoutTests/editing/selection/focus-and-display-none-and-redisplay.html

<!DOCTYPE html>
<body>
<script src="../../resources/js-test.js"></script>
<textarea>original</textarea>
<script>
description('Unfocused TEXTAREA should accept editing.');
jsTestIsAsync = true;
var textarea = document.querySelector('textarea');
textarea.addEventListener('blur', function() {
    textarea.style.display = 'inline-block';
    shouldNotBe('document.activeElement', 'textarea');
    document.execCommand('inserttext', false, 'add');
    shouldBeEqualToString('textarea.value', 'addoriginal');
    finishJSTest();
}, false);
textarea.focus();
textarea.style.display = 'none';
</script>
</body>
